Squashed 'Externals/Core/' changes from d056cff..bd2ee59
bd2ee59 Long to int.

git-subtree-dir: Externals/Core
git-subtree-split: bd2ee597cdc607f61bdc31f5ca941e96b55ad2e0

+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Merchant/AcquiringOfferDto.cs
@@ -13,7 +13,7 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
         public Guid OfferGuid { get; set; }
 
         [DataMember]
-        public long SwitcherCurrentBankKey { get; set; }
+        public int SwitcherCurrentBankKey { get; set; }
 
         [DataMember]
         public int LocationTurnoverMonthly { get; set; }
@@ -61,7 +61,7 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
         public bool AMEXNewAccount { get; set; }
 
         [DataMember]
-        public long MCCKey { get; set; }
+        public int MCCKey { get; set; }
 
         [DataMember]
         public GenBusinessLegalTypeEnum BusinessLegalTypeKey { get; set; }
