+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Merchant/AddProspectDto.cs
@@ -8,7 +8,7 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
 
         public string LocatorId { get; set; }
 
-        public long MainPhoneCountryKey { get; set; }
+        public int MainPhoneCountryKey { get; set; }
 
         public string MainPhoneNumber { get; set; }
