+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Merchant/OfferMinimumMonthlyServiceChargeDto.cs
@@ -2,9 +2,12 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
 {
     public class OfferMinimumMonthlyServiceChargeDto
     {
-        public long MmscKey { get; set; }
+        public int MmscKey { get; set; }
+
         public string Name { get; set; }
+
         public string Description { get; set; }
+
         public double Value { get; set; }
     }
 }
